Workaround bad protocol while sending task expired exceptions
Summary:
Since the header is managed by HeaderServerChannel rather than
by the request, it is shared between multiple requests. Consider:

1. A binary request arrives. The request will take 100sec. Task expire
timeout is set to 5sec.
2. A short compact request arrives, is handled and reply returned. Due
to the header being shared, this also sets the protocol inside the
header to compact.
3. 5 sec pass and the first request times out. We need to send a
TApplicatationException("Task expired"). For that we need to get the
method name of the original request, and for that we need to deserialize
the first several fields in the request. We try to do that with compact
protocol instead of binary.

This diff workarounds the issue by trying the "other" protocol, if the
one from the header can't deserialize.

Test Plan:
Made a client that does steps 1-3 above. See a stack trace
similar to https://phabricator.fb.com/P16723523 before, see a normal
task expired response after.

Reviewed By: haijunz@fb.com

Subscribers: alandau, bmatheny, mshneer

FB internal diff: D1618573

Tasks: 5382924