Fallback to insecure (and not fatal) in case of kerberos config parse errors
Summary:
Modified Krb5Context ctor so that it doesn't fatal if context cannot be
initialized due to a bad config. Instead throw and exception.

Modified Krb5credentialsCacheManager ctor to catch these exceptions and ignore.
The ccache manager thread tries to initialize the context / store if they are
not already initialized. The exceptions are caught and propagated from the
thread in the same way.

Test Plan: unit tests

Tested manually by building a client locally which talks to a test service.

Reviewed By: andrei.bajenov@fb.com

Subscribers: trunkagent, alandau, bmatheny, mshneer, shamdor

FB internal diff: D1630742

Tasks: 4212569