Reduce the default number of threads for test servers.
Summary:
[Thrift] Reduce the default number of threads for test servers.

Reduce to 1 worker thread. The reason is that, in test cases, we usually just want a convenient way to run a server in the background and send it one or two requests. We mostly don't need to do stress-testing in our unit/module-level tests. In my trials, this trims the taken time per test that but instantiates a `ScopedServerInterfaceThread` from 45-50ms to 25-30ms.

Provide an alternative constructor for people who want to customize their own `ThriftServer`. It's probably not terribly needed since basically ends up being a wrapper around `ScopedServerThread`, but, for consistency.

Test Plan: Unit tests in `thrift/lib/cpp2/test/util/ScopedServerInterfaceThreadTest.cpp` and continuous build.

Reviewed By: njormrod@fb.com

Subscribers: alandau, bmatheny, mshneer, dougw

FB internal diff: D1631258