Fix the thrift server behavior when security kill switch is removed / turned off
Summary:
Updated the logic which processes the kill-switch while creating a
connection. Make sure that when kill switch is enabled, a thrift server running
on the host:
- Doesn't upgrade to permitted from disabled.
- Downgrades to permitted from required.
- Doesn't change if it was already permitted.

Make sure that the behavior of the server reverts to the same as specified by
the sasl_policy once the kill switch expires or is turned off.

Improve server logs by logging the exception if SASL fallback to insecure
fails.

Test Plan: unit tests

Reviewed By: mhorowitz@fb.com

Subscribers: trunkagent, alandau, bmatheny, mshneer, shamdor

FB internal diff: D1635557

Tasks: 5448400