remove ThreadGate from thrift
Summary: `GenericThreadGate` was being (ab)used in a thrift test, and we had some code generation for a `future_gate` variant, but it doesn't seem to be used anywhere (and given the lack of an include of appropriate header files, or tests, probably wouldn't work anyway).

Test Plan:
I built and ran the edited test.

Reviewed By: davejwatson@fb.com

Subscribers: trunkagent, net-systems@, ruibalp, fugalh, alandau, bmatheny, exa, mshneer

FB internal diff: D1637557

Tasks: 5409538