handle nulls in SimpleJSONProtocol
Summary:
In a struct serialization like `{"foo": 5, "bar": null}`, we should treat that the same as if the bar member wasn't declared at all (`{"foo": 5}`).

To do this, we peek ahead after reading a field, and see if the value is null.  If it is, we skip over it and read the next field.

Test Plan: g-unittest

Reviewed By: haijunz@fb.com

Subscribers: trunkagent, alandau, bmatheny, mshneer

FB internal diff: D1674617

Signature: t1:1674617:1416332891:bdbc6d04731a6d66de537a60b5a24667878d49d0