Encrypt in-place -- avoid unnecessary memcpy
Summary:
This diff makes the wrap / encrypt path be in-place for secure thrift. It eliminates two buffer copies:
- one during coalesce
- one during buffer allocation inside krb5 code

This is about 14% perf boost seen on loadgen with 100kb requests.

The decrypt path is a little more complicated, I'll try to see if it's possible to do it in-place given the current apis tomorrow. If possible, I'll do that in a separate diff. The complication is around knowing the padding size in advance.

Test Plan: loadgen, unittests

Reviewed By: alandau@fb.com

Subscribers: hannesr, trunkagent, alandau, bmatheny, mshneer, mhorowitz, acampi, wstefancik

FB internal diff: D1733275

Signature: t1:1733275:1418428637:fa3b39aad33d95bd3f300613e3073793e6e0388c