protect ThreadManager with a lock
Summary:
The handler may call ThriftServer::getThreadManager on its background threads
before ThreadManager is created

There's a race in ServiceFramework which may lead to a segfault on startup:
1. Create a handler
2. Create ThriftServer without ThreadManager
3. Call ServiceFramework::addThriftService on the handler
4. Call ServiceFramework::go

This could lead to a segfault if getCounters is called between steps
3 and 4 (could happen on handler's background threads) before ThriftServer
creates a default ThreadManager.

In ZippyDB, one of ThriftServers is created without
ThreadManager -
https://phabricator.fb.com/diffusion/FBCODE/browse/master/zippydb/server/RocksThrift2Server.h;7b38377054ec256df0847c72cc63277385d24b23$165-169).
Rather than giving it an explicit ThreadManager, I think it's better to fix
this in Thrift/ServiceFramework.

Test Plan: ran ZippyDB integration test in a loop without segfault

Reviewed By: davejwatson@fb.com

Subscribers: fbcode-common-diffs@, davejwatson, andrewcox, alandau, bmatheny, mshneer, yfeldblum

FB internal diff: D1821082