return error codes in headers
Summary:
If srproxy encounters a TTransportException or TServiceRouterException when sending the request to the server, we return a TApplicationException, and we lose the error codes.  Let's set those in the headers.

In the python srproxy client we can hack the protocol to throw the correct error code if the headers are set.

Test Plan: g-unittest

Reviewed By: mshneer@fb.com

Subscribers: trunkagent, mcduff, hitesh, alandau, bmatheny, mshneer, yfeldblum

FB internal diff: D1856508

Signature: t1:1856508:1424373150:8ca90768bbf6cebef95db7d539b1d10b22f8ee61