Fix bug when skipping bool in fastproto
Summary:
In compact readBool can return 0 so skip isn't failing in this
case.

Test Plan:
Caught this bug when trying to replace
thrift/test/py/SerializationTest.py with fastproto in D1868315.
testBackwards fails when the bool field in a VersioningTestV2 object
needs to be skipped because it is deserializd as a VersioningTestV1
object. With this fix it works.

Reviewed By: alandau@fb.com

Subscribers: ruibalp, alandau, bmatheny, mshneer, yfeldblum

FB internal diff: D1869499

Signature: t1:1869499:1424830344:b7eb53aead2ff26e41c0f9446339bda7b6be6b53