No-handshake Kerberos negotiation
Summary:
This adds support for the no-handshake kerberos negotiation. Basically the client sends a token to the server followed by an encrypted request.

Even though this is not technically "mutual authentication", since the requests are encrypted and authenticated from the client, we're not losing much in terms of security. Sure, the client doesn't know that it's sending the encrypted data to the right server, but if it's not, then the server can't do anything with it anyway.

The plan in the next diff is to allow servicerouter to set the type of negotiation. That will be controlled via an smc flag. We default to the 1 round trip handshake.

Note that if the server fails to decrypt the request, in this case there is no "fallback". It will be just a failed request. This is because at this point the server would have lost all framing info. I guess we could kill this channel and retry the request without security... but that would need to be done at the servicerouter level.

Lastly, because we're not negotiating the encryption type in this mode, we default to the key's encryption type. Since the keys are not GCM, we're not using GCM mode.

Test Plan: I hardcoded different values of the mech, and made sure they all worked. I tested using loadgen and a simple client.

Reviewed By: alandau@fb.com

Subscribers: twt, saw, davejwatson, alandau, bmatheny, mshneer, yfeldblum, srenfro, mhorowitz, wstefancik

FB internal diff: D1977425

Signature: t1:1977425:1429035895:dc797d4d8244c87c5fa1349c0e3b43d8c2b89868