Fix RequestContext flow for security establishment
Summary:
This is a common pattern in our code base where multiple requests
(each with *different* RequestContext) would queue up for some work to finish.
After that work is finished, callback of each request is invoked with wrong
*same* RequestContext. The RequestContext would be of the first request queued.

We are fixing this pattern by saving RC and restoring it back when callbacks
are invoked. This is similar to SR selection fix in D1914213.

This is a common pattern and we can't rely on fixing all the present and future
use cases in fbcode. This whole RequestContext thing seems fragile to me :(

Test Plan: Added unit test which was failing without this fix.

Reviewed By: davejwatson@fb.com

Subscribers: alandau, bmatheny, yfeldblum, sandeepkk, andreib, andrewcox

FB internal diff: D2012091

Tasks: 6337419

Signature: t1:2012091:1429725958:e7257748310c8bf38aeeb0a47f34ee75d9d796fc