Modify thrift compiler to generate cache invalidation keys
Summary:
The thrift compiler changes allow for reading invalidation keys from
request buffers.

This modifies the thrift2 compiler to parse an annotation
"cpp[2].cache" on arguments to methods. This information
is then used to generate a map of method name to cache key.

The cacheKey is a parameter in the method signature that is
used as the basis of what the thrift request and response
cache index for invalidation should be.

This information is used in the generated method getCacheKey
which for an input buffer will return the value of the
cacheKey parameter (if it exists, otherwise an empty Optional).

This can then be called from a custom processor to determine
whether or not the request should be cached

Test Plan:
fbconfig thrift/lib/cpp2/test/
fbmake dbg;
_bin/thrift/lib/cpp2/test/ThriftServerTest
--gtest_filter=*CacheAnnotation*

Reviewed By: ckwalsh@fb.com

Subscribers: alandau, bmatheny, yfeldblum

FB internal diff: D2020015

Signature: t1:2020015:1429848466:6b35ca61bae1eb8ac1ada3f8bfc6cbf82d50c54a