thrift + lua
Summary:
This code has not had a non-codemod diff since 2012. The code does not appear to be used by anyone (verify with fbgs thrift/lib/lua and fbgs thrift/test/lua). The code makes use of lua_*_deprecated, and therefore appears to be deprecated code. We are presently trying to eliminate the last of the lua_*_deprecated build rules so that we can upgrade everyone to Test Pilot.

If this is really something we want to keep, I will go through and attempt to fix up the modules, but if nobody is invested in this, I'd rather just kill it.

Test Plan: Fire.

Reviewed By: alandau@fb.com

Subscribers: woo, pgriess, davejwatson, afrind, haijunz, njormrod, ps, bmatheny, yfeldblum, ldbrandy, tudorb, alandau, andrewcox

FB internal diff: D2034593

Tasks: 6610527, 6486445

Signature: t1:2034593:1430415129:3fcef959c51f211a4fc812c21897969d13b1fa83