Fix fastproto handling binary data
Summary:
Use the correct folly::StringPiece (a.k.a Range<const char*>
constructor. The one that accepts a char* uses strlen() to calculate the
length. This is incorrect for binary data.

Test Plan:
In #6997880 the binary data passed to fastproto starts with a
null byte so it is written as an empty string. With this it works. Also
fixed fastproto unit test to test with binary data starting with null
bytes and without this change it fails.

Reviewed By: dpittman@fb.com

Subscribers: ruibalp, alandau, bmatheny, yfeldblum, dpittman, maxmg

FB internal diff: D2053124

Tasks: 6997880

Signature: t1:2053124:1431021024:ffca7971ba87493d4816af68dae5f316316a6063