Allow setting IO exe with a preset number of threads
Summary:
Several setters on the ThriftServer used CHECK(ioThreadPool_->numThreads() == 0) as a proxy to determine if the
server is running and whether it is safe to modify its config. This in turn prohibits setting the IO thread pool with a preset number of threads.

This diff replaces these checks with a proper check using a flag which indicates whether it is safe to modify the server config.

Test Plan:
Unit tests and checked the use of setIOThreadPool (which no one else
seems to do).

Reviewed By: davejwatson@fb.com

Subscribers: fugalh, folly-diffs@, chalfant, andrewcox, trunkagent, alandau, bmatheny, yfeldblum

FB internal diff: D2034584

Signature: t1:2034584:1432044921:f81f6e1d27d7e113c9ef862f41617607bb7ae77d