Fix generated code for struct with a cpp.ref union field
Summary:
In cpp1, if a field is a reference to a struct, the generated code
checks if any of its field is set. This doesn't work for union which
doesn't have isset method. Change to test if the union's type is set.

This doesn't exist in cpp2 because of D2141057.

Test Plan:
Generate and build code for such a struct, visually check the
generated code.

Reviewed By: alandau@fb.com

Subscribers: trunkagent, alandau, yfeldblum

FB internal diff: D2148105

Tasks: 7328445

Signature: t1:2148105:1434064996:d2ca74b07c8f522359786c1b2d0c4471c102cf65