Fix Test: FutureGetOrderTest has too small a queue.
Summary:
[Thrift] Fix Test: FutureGetOrderTest has too small a queue.

The test frequently fails because its thread-manager queue is too small for the test case.

The test case immediately issues 5 requests but the thread-manager queue has space only for 2 requests. If the thrift-server in any given run enqueues all 5 requests to the thread-manager first before dequeueing any of them for execution, some of the requests will fail to be enqueued, an exception will be delivered to the client, and the test will fail.

Facebook:

Also removes a dependency on internal Facebook code, because this is OSS and that's just not cool.

Test Plan: Unit test: `thrift/lib/cpp2/test/FutureTest.cpp`.

Reviewers: alandau, davejwatson, mcduff, njormrod, haijunz

Reviewed By: haijunz

Subscribers: alandau, bmatheny, yfeldblum, dougw, brettp

Differential Revision: https://phabricator.fb.com/D2217879

Signature: t1:2217879:1436213639:efd39eeffeff7edbfb082d3906083958a283294f