Remove unneeded flaky tests for TNonblockingServer
Summary:
TNonblockingServer is deprecated in python for a while. We
still have a few tests that is broken or flaky for this server and the
tasks open for it are closed So no need to keep those tests around.

Test Plan: fbconfig thrift/test/py && fbmake runtests

Reviewers: andrewcox, yfeldblum, alandau

Reviewed By: yfeldblum, alandau

Subscribers: alandau, yfeldblum

Differential Revision: https://phabricator.fb.com/D2228100

Signature: t1:2228100:1436393063:5b8b0285cf81d76a232b69655a6f9daf77ac9622