A test case covering inheritance with cpp2 processors.
Summary:
[Thrift] A test case covering inheritance with cpp2 processors.

This test case is a common scenario of having two thrift service, one inheriting the other, and verifying that when we run a server with the derived service, the derived processor correctly handles messages sent to it corresponding to methods of the base service. And, overall, that cpp2 clients talking to derived-service cpp2 servers just work. This scenario is common in practice in our codebase and so it deserves coverage in the thrift unit-tests.

Test Plan: Unit tests: `thrift/lib/cpp2/test/ThriftServerInheritanceTest.cpp`.

Reviewers: alandau, davejwatson, mcduff, njormrod, haijunz

Reviewed By: haijunz

Subscribers: alandau, bmatheny, dougw, yfeldblum, brettp

Differential Revision: https://phabricator.fb.com/D2229573

Signature: t1:2229573:1436461970:0102ed4a88d9332618b7ba3f5bb5e8e9a8308f9a