Fix cpp2 server code for EventHandlers that rethrow.
Summary:
[Thrift] Fix cpp2 server code for EventHandlers that rethrow.

Before D2219542, exceptions from the generated handler deriving the `SvIf` are delivered to the `HandlerCallback` using `std::exception_ptr`.

Even exceptions that are never actually thrown by the deriving class, but which are instead delivered via the `future_` method, are rethrown by the generated handler's `async_tm_` or `async_eb_` methods to be delivered to the callback using `std::exception_ptr`. It is also possible to deliver exceptions to the callback using `folly::exception_wrapper`, but that is not how the generated code does it actually does it.

Whichever way exceptions are delivered to the callback, that is the way the callback delivers exceptions to the `GeneratedAsyncProcessor`'s generated `throw_` or `throw_wrapped_` methods.

The processor's `throw_` method rethrows the exception and, in the catch block, calls the `ContextStack`'s and through that the `TProcessorEventHandlers`' `userException` or `handlerReady` methods. These methods are therefore called with there being a `currently handled exception` in existence: `throw;` and `std::rethrow_exception(std::current_exception());` would work. This behavior looks like an implementation detail. But some event handlers do rely on it, and the blamed diff causes them problems.

Note that the processor's `throw_wrapped_` method does not rethrow the exception, so the event handler will not see a `currently handled exception` anywhere. But the generated `async_tm_` or `async_eb_` methods will not actually call this.

D2219542 switches the generated `async_tm_` and `async_eb_` methods to delivering exceptions to the callback using `folly::exception_wrapper`. So when event handlers are notified of exceptions, there is no `currently handled exception` in existence, and their reliance on that implementation detail breaks them.

But some event handlers do need access to the exception, so we temporarily revert to this implementation detail. This diff delivers exceptions to the callback using `std::exception_ptr`. However, in the future we may switch to deliverying exceptions only as `folly::exception_wrapper` while supporting another way of delivering exceptions to the event handlers.

Facebook:

Causes `ProxyTestFixture.testWithProxying` in `payment/libPayment/test/ProxyTest.cpp` to pass once again.

Test Plan: Unit tests and CI.

Reviewers: alandau, davejwatson, mcduff, njormrod, haijunz

Reviewed By: haijunz

Subscribers: upashkevich, alandau, bmatheny, yfeldblum, taowei, sushiln, shweng, kph, dougw, brettp

Differential Revision: https://phabricator.fb.com/D2233217

Tasks: 7627108, 7632835, 7627104

Signature: t1:2233217:1436548540:926b00499a0c3d9a03971a831b3070f364c1ed8c

Blame Revision: D2219542