thrift fuzzer: fix exception gathering for non-consecutively-numbered throws
Summary:
Exception lists for functions can have gaps in numbering, e.g.:

  void foo () throws {
    1: ExceptionOne eo,
    3: ExceptionThree et
  }

In this case, there is a `None` in the result thrift_spec tuple. `None`s need to be skipped when gathering exceptions that a method throws.

Test Plan: Run the fuzzer on a service with a function with non-consecutive numbering in the throws block, e.g., cachius

Reviewers: haijunz, alandau, mcduff, yfeldblum

Reviewed By: yfeldblum

Subscribers: alandau, bmatheny, yfeldblum

Differential Revision: https://phabricator.fb.com/D2238547

Signature: t1:2238547:1436828466:a70fb49eee3074f15dc5a236f1291944caf73bcb