Extract a cpp2 client test.
Summary:
[Thrift] Extract a cpp2 client test.

Also change the server part to run entirely in the `EventBase`. The test was failing for me infrequently, but not never, because it took > 1ms to dequeue the task in a `ThreadManager` thread, returning a `TApplicationException` with type `TIMEOUT` to the client instead of the `TTransportException` with type `TIMED_OUT` that the test is expecting. The difference is that the test is trying to test client timeouts, not server timeouts, so let's not issue any server timeouts.

Test Plan: Unit tests: `thrift/lib/cpp2/test/ThriftClientTest.cpp`.

Reviewers: alandau, davejwatson, mcduff, njormrod, haijunz

Reviewed By: haijunz

Subscribers: fugalh, jsedgwick, hannesr, alandau, bmatheny, yfeldblum, dougw, brettp

Differential Revision: https://phabricator.fb.com/D2235801

Signature: t1:2235801:1436894847:abf37c775506fa050aea3bc4b1c2f2f65d8588b1