Skeleton for accepting login service tokens
Summary:
This basically adds skeleton code to start accepting login service tokens in addition to kerberos credentials. Unfortunately this makes the makeAclQuery call somewhat more expensive because it's now doing RSA verification.

I tested this on the crypto service (where we do an ACL check on every request). I ran a load generator against it. The cpu before the diff was 600%, after the diff it was 1000%. I'm not sure if we could do better though. Maybe in the future we can cache verification results, but that's probably not a good idea for a v0.

Test Plan: Passed a legit token in the headers, made sure client_uid is set correctly.

Reviewed By: slarsen@fb.com

Subscribers: trunkagent, alandau, bmatheny, yfeldblum, micha, zhengmi

FB internal diff: D2267140

Signature: t1:2267140:1437672504:610f74be653e63c63efcbae7f7235a355a6df18f