Capture things from handler callback
Summary:
This reverts D2256310. The issue there seems to be in the lambda
running on evb it accesses some member variables from the callback but the
callback is destructed immediately after the call to exception(). I saw
from the core dump protoSeqId is some random value.

Another way to fix it is release the callback ptr and call
resultInThread and exceptionInThread. The future fix itself is also
fine, just want to verify this is the problem.

Test Plan: Not tested yet, probably needs @abhinavbatra to canary it.

Reviewers: alandau, andrewcox, abhinavbatra

Reviewed By: abhinavbatra

Subscribers: vshubin, pweng, trunkagent, alandau, bmatheny, yfeldblum, abhinavbatra

Differential Revision: https://phabricator.fb.com/D2261948

Tasks: 7743550

Signature: t1:2261948:1437514913:1d1e4aebea7eea2b75fe4ed4d7dc96c72592fa05