thrift fuzzer: Type constraints
Summary:
Constraints can be specified for all values of a given type, regardless of which method, argument, or field they appear in.

To create a type constraint, prepend `|` to the type name and use it as a key in a constraints object. This can be at the top level constraints object (to apply to all values of the given type,) or under any other constraints object to only apply within that scope.

Examples:

  name=config.json
  constraints = {
    "|i32": {"choices": [1, 2, 3]},
    "|MyStruct.x.choices": ["a", "b"],
    "my_method.arg1|float": {"mean": 3} // Applies to all floats that are subvalues of arg1
  }

This is especially useful for structs that are reused. For example, `titan/iris` has a `MessageMetadata` struct that is a subfield of 12 other struct types.

Depends on D2241191

Test Plan: fbconfig thrift/test/fuzzer && fbmake runtests

Reviewers: haijunz, mshneer, yfeldblum

Reviewed By: yfeldblum

Subscribers: alandau, bmatheny, andrewcox, sjenkins, yfeldblum

Differential Revision: https://phabricator.fb.com/D2247928

Tasks: 7527791

Signature: t1:2247928:1437607419:20966ba8cd85a9dbda3506fcd2678cba6a16ef2f