Add new fixture for Android
Summary:
This diff adds a fixture for the Android compiler to demonstrate the changes in D2285027. We simply just re-use the existing fixture that was already written for Haskell.

This generated code is //wrong// and this commit only serves to show why D2285027 is //right//. The code will compile but if actually run will fail because the value of `mMap.get(Module.MyStruct_MyIncludedField))` in ModuleLogger.java:42 is cast to a `ModuleLogger` when it really is an `IncludesLogger`.

Test Plan: These are fixtures -- I ran `thrift/compiler/test/build_fixtures _bin` to make them.

Reviewed By: yfeldblum@fb.com

Subscribers: alandau, yfeldblum, pedrodemasi

FB internal diff: D2288314

Signature: t1:2288314:1438117612:a9043d4932ae3135d00c54088ebc97f3e25256d6