Fix issues with including external thrift structs
Summary:
I found a bug in the Android compiler with regards to structs included from different thrift source files. Apparently, I hadn't tested it as rigorously as I had hoped.

Essentially if a struct foo.Example included a struct bar.Baz within it, when the Android code tried to `write` the Baz field, it would erroneously cast it to a `FooLogger` instead of a `BarLogger`. We fully-specify all of the package names every time we cast to an imported struct (which as far as I can tell is the same behavior as the Java compiler).

Test Plan:
* Made two thrift files and had one include structs from the other
* Compiled the android_lite code.
* Made sure `javac` accepted it

* Added the appropriate fixture. Check them out below.

Reviewed By: yfeldblum@fb.com

Subscribers: alandau, yfeldblum, pedrodemasi

FB internal diff: D2285027

Signature: t1:2285027:1438295829:5a3991a3edee3e2c48917a0df930614592b9b6ee