don't monkeypatch select so select still has epoll
Summary:
when gevent monkey patches select, it does not expose an
'epoll' member function. D2283795 added a thrift tgeventserver to
fbossdeploy, but this broke things in weird ways because we were
monkey patching select. Basically, fbpackage uses fbpkg.torrent
which uses a third-party tool called trollius for asynchronous
io. That library determines what type of selector to use based
on whether epoll, poll, kqueue etc. are in the select module. The
monkey patched select has none, so it defaults to a selector based
on select.select, which also happens to be broken because in tp2,
because it is never used. I separated this into a separate diff so
other users of TGeventServer know, but it looks like it is used in
bgpd and a few experimental libs only.

Test Plan:
Fbossdeploy building an ephemeral package works now and
the thrift server still performs as expected.

Reviewed By: cooper@fb.com

Subscribers: net-systems@, alandau, bmatheny, yfeldblum, yogeshwer

FB internal diff: D2308975

Signature: t1:2308975:1438660592:9367d977f434dc6876323b794dd52fa47bbdc11d