header per request
Summary:
Currently client and server channel own a THeader object and
use it for all requests. Some states managed in THeader are shared by
all requests, while some are specific to one request. So the idea here
is to manage the common states in a higher level and let request owns
its own THeader.

In HeaderClientChannel, A THeader is created before sending a request
and common states like transforms, client type, flags are copied from
the channel. Once the request is sent this THeader is destroyed. When
receiving a response a THeader is created at removeFrame and passed to
the TwowayCallback, then it is moved into a ClientReceiveState and
passed to the user supplied RequestCallback. If it is a sync call, the readHeaders
will be moved to the RpcOptions which is passed in when making the call.

In HeaderServerChannel, calling removeFrame creates a THeader as well.
This THeader is then owned by the HeaderRequest object and a pointer of
it is passed to the Cpp2RequestContext. Since this header is used
exclusively by the request it is safe to use it in the context to check
and add headers. The Cpp2RequestContext still copies some states from
the channel that may be changed, like minCompressBytes. This THeader has
the same lifecyle of the request and is also used for sending responses.

In service router, a THeader is passed in when calling sendRequest. write headers
are still stored in the Thrift2RequestContext because this request might be retried.
For shadow requests, a new THeader is created for each of them. When receiving,
the THeader passed from HeaderClientChannel will be available from ClientReceiveState,
just like in HeaderClientChannel.

Test Plan: Thrift and servicerouter unit tests, check unit tests of services affected. Canary on hero.dark tier

Reviewed By: andrewcox@fb.com

Subscribers: net-systems-diffs@, trunkagent, zeus-diffs@, jasont, net-systems@, tw-eng@, kennyyu, raghavpi, chaoyc, vikas, vloh, hero-diffs@, tingy, sergioclemente, dihde, search-fbcode-diffs@, unicorn-diffs@, benj, maxwellsayles, yzhan, sdoroshenko, jsedgwick, fbcode-common-diffs@, davejwatson, andrewcox, mcduff, hitesh, alandau, bmatheny, wormhole-diffs@, mshneer, jmkaldor, yfeldblum, jhunt, haijunz

FB internal diff: D1984738

Signature: t1:1984738:1438908897:b5b02b1ae7fd3abc759b3b1618b2c34a99a9313d