Use parsed.netloc for the HTTP Host header
Summary:
ParseResult.host normalizes IPv6 addresses in URIs to not
include the square braces, but it is illegal to pass an IPv6 IP address
without the braces in the HTTP Host: header.

Fix this be remembering the .netloc attribute of the ParseResult and
using that instead of the host we pass onto the http client.

Test Plan: thrift unittests

Reviewed By: haijunz@fb.com

Subscribers: wch, alandau, bmatheny, yfeldblum, arunm

FB internal diff: D2335654

Signature: t1:2335654:1439366063:6a6744f78389842a28a967db5a7ec4be6fdc63d0