Fix sendCatchupRequest should not move unique_ptr of THeader
Summary:
If this is moved, it will be destroyed after the response is
sent. If it is a cpp1 client which doesn't use out of order, timeout fires
and a response is sent, the THeader is destroyed. Later when the real
response is sent it will see a nullptr.

Test Plan: thrift cpp2 unit tests pass.

Reviewed By: alandau@fb.com

Subscribers: arafay, alandau, bmatheny, yfeldblum

FB internal diff: D2386391

Tasks: 8136456

Signature: t1:2386391:1440634487:eace8b404375d0afa7f1e5a23228998b6ecebccc