Perf: Possibly deliver exceptions to TProcessorEventHandler without rethrowing.
Summary:
[Thrift] Perf: Possibly deliver exceptions to TProcessorEventHandler without rethrowing.

Exceptions from the generated handler deriving the `SvIf` are delivered to the `HandlerCallback` using `std::exception_ptr`.

Even exceptions that are never actually thrown by the deriving class, but which are instead delivered via the `future_` method, are rethrown by the `HandlerCallback::complete()` family of methods, captured in `std::exception_ptr`, and passed to the `HandlerCallback::exception()` family of methods. If `HandlerCallback::exception()` gets `std::exception_ptr`, then the callback delivers exceptions to the `GeneratedAsyncProcessor`'s generated `throw_` methods; if it gets `folly::exception_wrapper`, it delivers exceptions to the `throw_wrapped_` methods.

The processor's `throw_` method rethrows the exception and, in the `catch` block, calls the `ContextStack`'s and through that the `TProcessorEventHandler`s' `userException` or `handlerReady` methods. These methods are therefore called with there being a `currently handled exception` in existence: `throw;` and `std::rethrow_exception(std::current_exception());` would work. This behavior looks like an implementation detail. But some event handlers do rely on it.

Note that, if the exception type is marked in the service method signature and the exception is delivered without it being thrown, the processor's `throw_wrapped_` method does not rethrow the exception, and so the event handler will not see a `currently handled exception` anywhere. If the exception type is not marked or it is thrown and caught, however, `throw_wrapped_` will rethrow the exception to find the runtime type of the exception. But `HandlerCallback::complete()` does not actually call this.

In this diff, we introduce `TProcessorEventHandler::handlerErrorWrapped` and `TProcessorEventHandler::userExceptionWrapped` which take `folly::exception_wrapper` params, and change `HandlerCallback::complete()` no longer to rethrow. We now always call the processor's `throw_wrapped_` methods instead of `throw_`, and no longer rethrow exceptions that are marked in the service method signature and that are delivered without being thrown. We deliver exceptions to `handlerErrorWrapped` and `userExceptionWrapped`, if possible, without rethrowing and without leaving a `currently handled exception` in existence.

Also:
* Make the Payments proxy handling compatible, verifying that `ProxyTestFixture.testWithProxying` in `payment/libPayment/test/ProxyTest.cpp` continues to pass.
* Keep the ServiceRouter tests passing, verifything that `ApplicationException.Basic` and `ApplicationException.BasicWithAppErrorWhat` in `servicerouter/client/cpp2/test/ServiceRouterClientChannelTest.cpp` continue to pass. These set expectations on the precise content of the strings passed to `TProcessorEventHandler::userException`.

Test Plan: Unit tests, in particular `thrift/lib/cpp2/test/ThriftServerExceptionTest.cpp`, and CI.

Reviewed By: haijunz@fb.com

Subscribers: antonl, dcapel, upashkevich, subodh, ldbrandy, sjenkins, vloh, kph, alandau, dihde, bmatheny, dougw, yfeldblum, brettp

FB internal diff: D2226001

Tasks: 8168675

Signature: t1:2226001:1441955851:a774db0d161f45b167b7e8e70677aea0503e3e94

Blame Revision: D2219542, D2233217