Fix O(n^2) parsing of large HTTP responses
Summary:
When data is available from an http stream, we do an O(n) parse of the
message and if it is incomplete, await new data. Previously we would wait for 1
byte, but if we have parsed the Content-Length header we can be smarter, and
only parse again when we have the entire message.

Test Plan:
Call a thrift service that returns 200mb. It completes in a minute
rather than spinning for 3 hours.

Reviewed By: andrewcox@fb.com

Subscribers: alandau, yfeldblum, cschoepf

FB internal diff: D2447506