Deprecate setWriteHeadersThroughChannel
Summary:
Some clients uses the channel to set headers so this api is added
in header-per-request to maintain backward compatibility. It is weird because
clients created using this api will send headers only on its first request.
Changed to use persistent headers and remove this api.

I checked a few places such clients are used and they are only used to send one
request. But it is still a bad way to send headers.

Test Plan: unit tests in affected projects

Reviewed By: alandau@fb.com

Subscribers: fbcode-common-diffs@, tw-eng@, alandau, kennyyu, yfeldblum, sergioclemente, bmatheny, anob

FB internal diff: D2472170