Prevent TNonThrowingSimpleServer from throwing
Summary:
We don't have any mechanism for a TServer to report that it failed to start. The main thread is doomed to wait for it forever. So this adds a `handleServeError` method to TServerEventHandler.

The `TNonThrowingSimpleServer` has been throwing exceptions. Technically, since the exception was thrown on a `std::thread` that wasn't handling it, `std::terminate` was getting called, which is even worse. This diff fixes the code that tries to ferry exceptions from the server thread back to the main thread. It also fixes the communication so that the main thread always gets notified when the server failed to start.

This had manifested itself as sporadically failing unit tests and GreenWarden tasks, because the symptoms only appeared when there was a race to acquire a free network port.

This may not be kosher, since it changes `apache::thrift::TServer`. I wanted to get an opinion on that before I made any assumptions.

Test Plan:
fbmake runtests_dbg --stress-runs 48 'servicerouter/selection/tests:health_check2_test - Comprehension.TestFB303HeaderIPv6Http'

Reviewed By: ckwalsh@fb.com

Subscribers: ckwalsh, #servicerouter, alandau, yfeldblum

FB internal diff: D2535680

Tasks: 8596898, 8596895, 8596899, 6671156, 8596892, 8596894

Signature: t1:2535680:1444773171:a41cba7856663b89562a4becafd2878097fd97b7